EE-197: Move uint mappings into sub-module
